[PrimeChecker.py] Add PrimeChecker class
This class takes in a number as a string argument and checks whether it is
a prime number. The class has an 'is_prime()' function that first checks
if the string is empty, to which it returns False or if not the case, it
checks whether the number is greater than  1 (Since the least prime number
is 2). If the number has no remainder as it is divided by a range of
numbers from 2 to the number previous to the number in question, then we
return False (since prime numbers have only 2 factors, one and itself),
however, if not the case, that is our prime number and we return True